+++ projects/i-novation/ng-rapidforms/ng-bootstrap/src/public_api.ts
@@ -1,6 +1,6 @@
 // Material Rapidforms
 export * from './lib/rapidforms-ng-bootstrap.module';
 
-// Material
+// Bootstrap
 export * from './lib/templates/dynamicform/dynamicform-bootstrap-ng.component';
-export * from './lib/templates/dynamicform/dynamicform-bootstrap-ng.component';
+export * from './lib/templates/dynamicformelement/dynamicformelement-bootstrap-ng.component';