+++ projects/i-novation/ng-rapidforms/src/lib/models/validators/RegexValidator.ts
@@ -8,14 +8,14 @@ export class RegexValidator implements CustomValidator {
   message: string;
   rule: ValidatorFn;
   error: string;
-  regex: string;
+  regex: string | RegExp;
 
   getFormattedMessage() {
-    this.message = this.message.replace('{{regex}}', this.regex);
+    this.message = this.message.replace('{{regex}}', this.regex.toString());
     return this.message;
   }
 
-  constructor(regex: string, message?: string) {
+  constructor(regex: string | RegExp, message?: string) {
     this.regex = regex;
     this.message = message || '{{attribute}} has to match the regular expression \'{{regex}}\'';
     this.rule = Validators.pattern(this.regex);