+++ CHANGELOG.md
@@ -1,6 +1,36 @@
 # Changelog
 
-## 0.0.1.beta.2
+## 1.0.4-beta.0
+
+- make library angular 10 compatible
+
+## 1.0.3-beta.0
+
+- fix validation messages for number input
+- fix change handler call without provided function
+- fix number validator, disable group separator by default
+
+## 1.0.2-beta.0
+
+- add number input template for spectre, material and bootstrap 3
+- fix issue with number formatting in number input field
+- add disabled property to form control options
+
+## 1.0.1-beta.0
+
+- angular material: make build compatible with newest version of material
+- spectre theme: fix issue with checkbox element is-error class
+
+## 1.0.0-beta.1
+
+- fix peer dependencies issue
+
+## 1.0.0-beta.0
+
+- make library angular 9 compatible
+- add regexp type to regex validator
+
+## 0.0.1-beta.2
 
 - fix issue with selected option in options array
 - add changed property for functions to all elements